producer-service/requirements.txt
opencv-python==4.8.1.78
kafka-python==2.0.2
numpy==1.24.3 